refactor: Move drag-and-drop logic into DOM actions
- Extract the complex event handling logic from the FileTree.svelte
component into two reusable Svelte actions: draggable and droppable.

- "droppable" dispatches a clean "filesdropped" custom event upon a
successful drop that the component using the action can handle
appropriately (in this case, by moving a file).

This change significantly simplifies FileTree.svelte, making it more
declarative and focused on its core responsibility of rendering a file
node. It also provides reusable actions to easily implement
drag-and-drop in other parts of the app, should we need it.