fix(updater): Fix update modal changelog
Fixes a bug where we were incorrectly displaying the bundled (and
therefore current) changelog, rather than the new one associated with
the latest release.

This is achieved by embedding the full changelog in the updater's JSON
payload during the release workflow and having the client parse
it. The public GitHub Release notes are then edited to remain clean.